Rename database table names
The structure of tables is as following:

Timetable(origin, destination, date)
Train(code, departure, arrive, length, timetable_id)

A timetable is specify by an origin, destination and date. A timetable
is formed by a list of trains. Each train has a code, departure and
arrive time and length of trip.